requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jugaad-trader==0.20
nsepy==0.8
requests-cache==1.2.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
lark==1.3.1
mlflow==2.13.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
prometheus-client==0.20.0
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0